contract(feat): move eip domain fetcher to function
