+++ src/controllers/urls.ts
@@ -6,7 +6,7 @@ import * as TagControl from './tags';
 import * as UserControl from './users';
 import { UserDocument, TagDocument } from '../models/schema';
 import { Data } from '../types/typechain/Channel4';
-import { getContractObject } from './contract';
+import { getEIPDomain } from './contract';
 import { ethers } from 'ethers';
 
 
@@ -162,14 +162,7 @@ const getContentToSign = async (url: string): Promise<Data.ContentToLitigateStru
 };
 
 const getUrlEIP712Metadata = async () => {
-  const contract = getContractObject();
-  const EIP712Domain = await contract.eip712Domain();
-  const domain = {
-    name: EIP712Domain.name,
-    version: EIP712Domain.version,
-    chainId: EIP712Domain.chainId,
-    verifyingContract: EIP712Domain.verifyingContract,
-  };
+  const domain = await getEIPDomain();
   const types = {
     ContentToLitigate: [
       { name: 'title', type: 'string' },
